+++ src/pm/hydra/mpiexec/pmiserv_cb.c
@@ -289,7 +289,7 @@ static HYD_status control_cb(int fd, HYD_event_t events, void *userp)
         HYDU_MALLOC_OR_JUMP(proxy->pid, int *, proxy->proxy_process_count * sizeof(int), status);
         int count, closed;
         status = HYDU_sock_read(fd, (void *) proxy->pid,
-                                proxy->proxy_process_count * sizeof(int),
+                                (int) (proxy->proxy_process_count * sizeof(int)),
                                 &count, &closed, HYDU_SOCK_COMM_MSGWAIT);
         HYDU_ERR_POP(status, "unable to read status from proxy\n");
         HYDU_ASSERT(!closed, status);
@@ -313,7 +313,7 @@ static HYD_status control_cb(int fd, HYD_event_t events, void *userp)
                             status);
         int count, closed;
         status = HYDU_sock_read(fd, (void *) proxy->exit_status,
-                                proxy->proxy_process_count * sizeof(int), &count, &closed,
+                                (int) (proxy->proxy_process_count * sizeof(int)), &count, &closed,
                                 HYDU_SOCK_COMM_MSGWAIT);
         HYDU_ERR_POP(status, "unable to read status from proxy\n");
         HYDU_ASSERT(!closed, status);
@@ -421,7 +421,7 @@ static HYD_status control_cb(int fd, HYD_event_t events, void *userp)
                 segment = strtok(current_list, ",");
                 HYDU_ASSERT(segment != NULL, status);
                 do {
-                    value = strtol(segment, NULL, 10);
+                    value = (int) strtol(segment, NULL, 10);
                     if (value == terminated_rank) {
                         included = 1;
                         break;
