+++ src/pm/hydra/proxy/pmip_utils.c
@@ -41,7 +41,7 @@ static HYD_status control_port_fn(char *arg, char ***argv)
     HYDU_ERR_CHKANDJUMP(status, NULL == port, HYD_INTERNAL_ERROR, "port not provided\n");
     name = strtok(port, ":");
     HYD_pmcd_pmip.upstream.server_name = name ? MPL_strdup(name) : NULL;
-    HYD_pmcd_pmip.upstream.server_port = strtol(strtok(NULL, ":"), NULL, 10);
+    HYD_pmcd_pmip.upstream.server_port = (int) strtol(strtok(NULL, ":"), NULL, 10);
 
     (*argv)++;
